Use PA2 for debug output
Reusing SWCLK is hard to work with, so use PA2 instead. On the current
board, this is a stepper mode pin, so using it as TX works as long as
the stepper is not used. On a future revision, the PA2 will be routed to
the AUX header to be more easily usable.