Support childbus protocol
This adds support for the child side of the childbus protocol, so this
firmware can be controlled and queried from the mainboard. Most of this
code is taken verbatim from the childbus bootloader.

This adds a single dummy GET_LAST_STATUS command, the actual commands
are yet to be determined and implemented.

The following files were copied from ChildbusBootloader, commit 29f1e52
(PROTOCOL: Bump version to 2.1):

    cp ../../ChildbusBootloader/{BaseProtocol.cpp,BaseProtocol.h,Crc.h,Bus.h} src/
    cp ../../ChildbusBootloader/stm32/{Rs485.cpp,Reset.cpp} src/stm32/