IrSensor: Allow "print_measurement()' method to be declared
The precompile variable 'ENABLE_SERIAL' prevented the method to be declared and
when ENABLE_SERIAL is not used, the compiler gives an error saying that the
there is no matching declaration. Everything in the print_measurement method is
already disabled when ENABLE_SERIAL is used, so there is no need to remove the
delaration. This also makes it easier to work with, since the method does not
need the "if defined" check where ever print_measurement is used.