Introduce sensor control options
This commit add two extra sensor states `STOPPED` and `MEASURING`, which can set
using the extra parameter for the `SET_STATE` command. The LED's will be turned
on in the main loop when the measuring state is set and turned off when
stopped. A newly introduced `GET_AND_CLEAR_SENSOR_VALUES` command can be used
to request the detected average blockage and maximum blockage values from the
hopper board.