IrSensor: Apply calibration correction
The min_value for each pixel is instantly lowered, but if the actual
value stays above the min_value, the min_value is slowly raised again,
to compensate for material stuck on the sensor, and for material that is
seen to briefly lower the observed values (presumably due to
reflections).

To prevent raising the values too high, there is a budget that is fed by
the periodic querying of the host, and shared among all pixels.